orders, order_items, shipping_address and lgareas migration updated

+++ database/migrations/2020_10_26_153852_create_order_items_table.php
@@ -15,8 +15,8 @@ class CreateOrderItemsTable extends Migration
     {
         Schema::create('order_items', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id')->index();
-            $table->integer('product_id')->index();
+            $table->bigInteger('order_id')->index();
+            $table->bigInteger('product_id')->index();
             $table->string('product_name');
             $table->bigInteger('product_quantity');
             $table->float('product_price', 15, 2);
